refactor: PaymentConfirmWebRequest 클래스를 적절한 패키지로 이동하여 모듈 구조 개선
