+++ src/main/java/roomescape/reservation/controller/dto/CreateReservationWithPaymentWebRequest.java
@@ -1,7 +1,7 @@
 package roomescape.reservation.controller.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import roomescape.payment.PaymentConfirmWebRequest;
+import roomescape.payment.controller.dto.PaymentConfirmWebRequest;
 
 public record CreateReservationWithPaymentWebRequest(
         @JsonProperty("reservation") CreateReservationWebRequest createReservationWebRequest,
